requirements.txt
Django==2.1.1
django-extensions==2.0.6
django-model-utils==3.1.1
djangorestframework==3.8.2
ipython==6.5.0
PyGithub==1.39
social-auth-app-django==2.1.0
